refactor: rename sign-in page to login and update styles and imports
