+++ src/components/common/auth-button.tsx
@@ -14,8 +14,8 @@ function AuthButton({
   return (
     <button
       type="submit"
-      className={`w-full h-10.5 rounded-full bg-midnight hover:bg-sakura text-white font-semibold
-         cursor-pointer ${className}`}
+      className={`w-full h-10.5 rounded-full bg-midnight hover:bg-white hover:text-midnight hover:border-midnight hover:border transition text-white font-semibold
+        cursor-pointer ${className}`}
       onClick={onClick}
     >
       {icon}
